go/687.longest-univalue-path.go
package main

/*
 * @lc app=leetcode id=687 lang=golang
 *
 * [687] Longest Univalue Path
 *
 * https://leetcode.com/problems/longest-univalue-path/description/
 *
 * algorithms
 * Easy (34.04%)
 * Total Accepted:    62.1K
 * Total Submissions: 182.4K
 * Testcase Example:  '[5,4,5,1,1,5]'
 *
 * Given a binary tree, find the length of the longest path where each node in
 * the path has the same value. This path may or may not pass through the
 * root.
 *
 * The length of path between two nodes is represented by the number of edges
 * between them.
 *
 *
 *
 * Example 1:
 *
 * Input:
 *
 *
 * ⁠             5
 * ⁠            / \
 * ⁠           4   5
 * ⁠          / \   \
 * ⁠         1   1   5
 *
 *
 * Output: 2
 *
 *
 *
 * Example 2:
 *
 * Input:
 *
 *
 * ⁠             1
 * ⁠            / \
 * ⁠           4   5
 * ⁠          / \   \
 * ⁠         4   4   5
 *
 *
 * Output: 2
 *
 *
 *
 * Note: The given binary tree has not more than 10000 nodes. The height of the
 * tree is not more than 1000.
 *
 */
/**
 * Definition for a binary tree node.
 * type TreeNode struct {
 *     Val int
 *     Left *TreeNode
 *     Right *TreeNode
 * }
 */
func longestUnivaluePath(root *TreeNode) int {
	if root == nil {
		return 0
	}

	cur := sameVal(root.Left, root.Val) + sameVal(root.Right, root.Val)
	left := longestUnivaluePath(root.Left)
	right := longestUnivaluePath(root.Right)
	return max(cur, left, right)
}

// find the longest uni value path that start from given node
func sameVal(t *TreeNode, val int) int {
	if t == nil || t.Val != val {
		return 0
	}

	left := sameVal(t.Left, val) + 1
	right := sameVal(t.Right, val) + 1
	return max(left, right)
}
